+++ app/src/app/core/validator.ts
@@ -1,0 +1,15 @@
+const rule = {
+    password: {
+        minLength: 6
+    }
+};
+
+const regex = {
+    password: `.{${rule.password.minLength},}`,
+    phone: '^0[67]\\d{8}$'
+};
+
+export {
+    regex,
+    rule
+};